+++ docs/changelog.rst
@@ -1,12 +1,14 @@
 Changelog
 ~~~~~~~~~
 
-0.16.0 (April 2023)
--------------------
+0.16.0 (May 2023)
+-----------------
 
+* Updated pytket version requirement to 1.15.
 * cost function now takes the same kwargs as process_circuits
 * add check for the number of classical registers to the backend
-* Updated pytket version requirement to 1.14.
+* add ``get_partial_result`` method to ``QuantinuumBackend``.
+* add ``Rxxyyzz`` gate support.
 
 0.15.0 (April 2023)
 -------------------